Port to Sinatra app
We need this to be more than just a static site now (form submissions,
sessions for user registration and support tickets) so we are porting
this to a Ruby/Sinatra app.

The middleman erb templates are compatible with Sinatra's erb templating
so this is mostly a matter of renaming the templates and swapping out
the middleman headers for `content_for` blocks to set title variables.